Optimize protocol - merge phases
The protocol changes to run all three phases in parallel.

The balance proof gets an additional boolean exit flag. Exit requests
are now sent off-chain to the enclave and the enclave will generate a
final balance proof with the exit flag set to true.

The challenge protocol is now:
1. At the end of epoch n, user doesn't receive balance proof (or doesn't
receive exit proof after sending an exit tx).

2. During on-chain epoch n+1, user posts a challenge using his n-1 BP.
This emits a Challenged event.

3. The epoch parses the Challenged event and handles the user as if he
sent an off-chain exit request. So at the end of epoch n+1, an exit
proof for the user is generated.

4. During on-chain epoch n+2, the operator responds with the exit proof
of the user. This emits a ChallengeResponded event containing the exit
proof.

5.a If the operator fails to respond, the contract is frozen to epoch
n-1 and all users can exit with their n-1 BPs.

5.b Otherwise, during on-chain epoch n+3, the user can finally exit with
the n+1 EP posted by the operator during on-chain epoch n+2. The user
has to reassemble their exit proof from the ChallengeResponded event.