Fix challenge protocol for phase merge
The challenge protocol needed to be changed further due to the phase merge.

There was still the case missing when a user challenges the operator for
an epoch in which the user was already exited from the system but didn't
receive their exit proof. In this case, the operator cannot respond with
a balance proof of the epoch after the challenged epoch. For this reason,
the respondChallenge function now also accepts exit proofs from the
challenged epoch, not only the one after.

For this reason, challenging users that challenge because they didn't
receive an exit proof now have to subscribe to ChallengeResponded events
from both, the challenged epoch and the one after.