Change contracts logic to receive deposits via TokenHolders
Before this, deposits were also made on the Erdstall contract.
This new logic has the added benefit that individual tokens don't need
to be registered by the Owner of the Erdstall contract. The owner now
only registers the TokenHolders per token type once and then arbitrary
tokens of the same type can be deposited into the TokenHolder.