PerunArt: Allow multiple minters to be added with ctor and addMinter
Before, the minter had to be set at deploy-time. Now, the deployer, who
is the owner of the contract, can add minters later by calling addMinter.

This is helpful in scenarios where we might not want to redeploy the
NFT contract during tests but still be able to set a newly deployed
ERC721 Token Holder contract as a minter by the deployer.